Update main.cpp
Big update made in order to accommodate new additions today. 
Additions include Servo functionality and facilitation for the detection of the fans with a digital input pin. The reason for this change is due to the detection of the fans being possible entirely in a physical circuit, lessening the load of a large c++ file. 

Skeletons for movement and detection of aliens and building complete. Full architecture for FPGA detection is untested and thus not included yet. 

General structure for data sending added. The idea is to limit the amount of data transferred on a need to know basis. For example, data will not be repeatedly send to Command but only sent if there is a change. This is to lighten the load on the communication and hopefully avoid errors. Implementation to begin Friday 10th June.